// src/types/types.ts
export type vehicles ={
    id:string,
    name: string;
    brand: string;
    color: string;
    price: number;
    year: number;
    license_plate: string;

}

export type vehiclesInputDTO ={
    name: string;
    brand: string;
    color: string;
    price: number;
    year: number;
    license_plate: string;

}